Login Form
This is a simple login form implemented using AWT (Abstract Window Toolkit) in Java. It provides a basic graphical interface with fields for entering a username and password, along with a login button. Upon clicking the login button, you can add your own login logic to authenticate the user. This code can be used as a starting point to build a more robust login system in Java. Feel free to customize and enhance it as per your requirements.